# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/chats/": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create a new chat",
        "description": "Creates a new chat room between the current user and a list of target users.\n\n- **user_id_list**: A list of user_ids for the users to include.\n- **Constraints**: \n    - Max 10 users per chat (initially, you can add more later).\n    - You cannot create a chat with non-existent users.\n    - The creator is automatically added to the chat.",
        "operationId": "create_chat_api_v1_chats__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user list (duplicates or too many users)."
          },
          "401": {
            "description": "Authentication credentials were not provided or are invalid."
          },
          "404": {
            "description": "One or more target users do not exist."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/chats/{chat_id}": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send a message",
        "description": "Persist a message to the database for a specific chat.\n\nReturns the message details including the generated ULID and timestamp.",
        "operationId": "send_message_api_v1_chats__chat_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessage"
                }
              }
            }
          },
          "401": {
            "description": "Authentication credentials were not provided or are invalid."
          },
          "404": {
            "description": "Chat not found"
          },
          "403": {
            "description": "User is not a participant of this chat"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get chat history",
        "description": "Retrieves all messages for a specific chat. Messages are sorted by ULID (chronologically).",
        "operationId": "get_chat_history_api_v1_chats__chat_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication credentials were not provided or are invalid."
          },
          "404": {
            "description": "Chat not found"
          },
          "403": {
            "description": "User is not a participant of this chat"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete a chat",
        "description": "**Hard deletes** a chat and all its associated history.\n\n- This operation performs a 'soft' check immediately.\n- The actual deletion of thousands of messages happens in a **Background Task** \n  to prevent the API from hanging.",
        "operationId": "delete_chat_api_v1_chats__chat_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication credentials were not provided or are invalid."
          },
          "404": {
            "description": "Chat not found"
          },
          "403": {
            "description": "User is not a participant of this chat"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Register a new user",
        "description": "Registers a new user in the system.\n\n- **Validation**: Checks if the email is unique.\n- **Security**: The password is securely hashed (Argon2) before storage.\n- **Output**: Returns the generated User ID (ULID).",
        "operationId": "sign_up_api_v1_users__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "example": {
                  "detail": "A user with this email already exists."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/chats": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get my chats",
        "description": "Retrieves a list of Chat IDs that the currently logged-in user participates in.",
        "operationId": "get_my_chats_api_v1_users_me_chats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserChatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication credentials were not provided or are invalid."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Deactivate my account",
        "description": "**Soft delete** the current user.\n\n- Marks the user as 'tombstoned' in the database.\n- The user will no longer be able to log in.\n- Historical data (messages) is preserved for data integrity.",
        "operationId": "delete_me_api_v1_users_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication credentials were not provided or are invalid."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login / Get Access Token",
        "description": "**OAuth2 compatible token login.**\n\nThis endpoint exchanges user credentials for a JWT Access Token.\n\n- **username**: Must contain the user's **Email Address**.\n- **password**: The plain-text password.\n\nReturns a Bearer token to be used in the `Authorization` header for subsequent requests.",
        "operationId": "login_api_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect email or password, or user is inactive.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Incorrect email or password"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_login_api_v1_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_v1_auth_token_post"
      },
      "ChatCreateRequest": {
        "properties": {
          "user_id_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "User Id List"
          }
        },
        "type": "object",
        "required": [
          "user_id_list"
        ],
        "title": "ChatCreateRequest"
      },
      "ChatCreatedResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          }
        },
        "type": "object",
        "required": [
          "chat_id"
        ],
        "title": "ChatCreatedResponse"
      },
      "ChatHistoryResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatHistoryResponse"
      },
      "ChatMessage": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "ulid": {
            "type": "string",
            "title": "Ulid"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "ulid",
          "timestamp",
          "user_id",
          "content"
        ],
        "title": "ChatMessage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageSendRequest": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MessageSendRequest"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
      "UserChatsResponse": {
        "properties": {
          "chat_id_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chat Id List"
          }
        },
        "type": "object",
        "required": [
          "chat_id_list"
        ],
        "title": "UserChatsResponse"
      },
      "UserCreateRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserCreateRequest"
      },
      "UserResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}